refactor: extract order desc card
